ticket #114
git-svn-id: https://chaste.cs.ox.ac.uk/svn/chaste/trunk@1767 b5f69653-2bf5-0310-8201-b0dafb8e1405

+++ mesh/src/common/AbstractElement.hpp
@@ -132,7 +132,7 @@ public:
         return centroid/((double)(ELEMENT_DIM + 1));
     }
     
-    long GetNodeGlobalIndex(unsigned localIndex) const
+    unsigned GetNodeGlobalIndex(unsigned localIndex) const
     {
         assert((unsigned)localIndex < mNodes.size());
         return mNodes[localIndex]->GetIndex();
